v0.9.2 -> v0.9.3
- [change] lua wait time for button reclick
- [add] CMD and ENTRYPOINT args to kubernetes file for Splash
- [remove] resource requirements for Splash since GKE ignores them

+++ scraper/food_crawler/spiders/dh_spider.py
@@ -32,8 +32,8 @@ class DHSPIDER(scrapy.Spider):
     current_day = date.today().strftime('%A, %B %-d, %Y')
     meals_list = tuple()
     # Page selectors
-    wait_time = int(os.environ.get('SPIDER_WAIT_TIME'))
-    splash_timeout = int(os.environ.get('SPLASH_TIMEOUT'))
+    wait_time = float(os.environ.get('SPIDER_WAIT_TIME'))
+    splash_timeout = float(os.environ.get('SPLASH_TIMEOUT'))
     dining_hall_sel = {
         'South': 'tr.cbo_nn_unitsPrimaryRow:nth-child(5) > td:nth-child(1) > div:nth-child(1) > table:nth-child(1) > tbody:nth-child(1) > tr:nth-child(2) > td:nth-child(1) > a:nth-child(1)',
         'North': 'tr.cbo_nn_unitsAlternateRow:nth-child(2) > td:nth-child(1) > div:nth-child(1) > table:nth-child(1) > tbody:nth-child(1) > tr:nth-child(2) > td:nth-child(1) > a:nth-child(1)'
